neuroep/requirements.txt
brainflow>=5.12.0
PyQt6>=6.6.0
pyqtgraph>=0.13.3
psychopy>=2024.1.0
scipy>=1.12.0
numpy>=1.26.0
mne>=1.6.0
pylsl>=1.16.0
matplotlib>=3.8.0
pyserial>=3.5
reportlab>=4.0.0
